% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Generate alignments from sequences}
\usage{
align(sequence_files, method)
}
\arguments{
\item{sequence_files}{file paths to sequence files}

\item{method}{alignment method}
}
\value{
filepaths to alignments, character vector
}
\description{
Create alignments from sequence files using method specified.
}
