\name{alignment}
\alias{alignment}
\docType{data}
\title{
Example alignment
}
\description{
Example alignment object for demonstration purposes.
}
\usage{data("alignment")}
\format{
  A large matrix object containing fake DNA sequence.
}
\examples{
data(alignment)
print(alignment)
}
\keyword{datasets}