\name{alignment_list}
\alias{alignment_list}
\docType{data}
\title{
Example alignment_list
}
\description{
Example alignment_list object for demonstration purposes.
}
\usage{data("alignment_list")}
\format{
  A large matrix object containing fake DNA sequence.
}
\examples{
data(alignment_list)
print(alignment_list)
}
\keyword{datasets}