% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{alignment_read}
\alias{alignment_read}
\title{Read in sequence alignment(s)}
\usage{
alignment_read(flpths, format = c("guess", "clustal", "msf", "mase",
  "phylip", "fasta"))
}
\arguments{
\item{flpths}{File path(s) to alignment file(s)}

\item{format}{Sequnece file format. Default 'guess'. Character.}
}
\value{
alignment list
}
\description{
Return alignment list object from reading in sequences from
alignment file(s). Use \code{format} to specify alignment file type. If no
\code{format} given, the function will guess based on file extension.
}
\details{
Available file formats: "clustal", "msf", "mase", "phylip" and 
"fasta".
}
\seealso{
\link{\code{seqinr::read.alignment}}
}
