% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{.alignment_read}
\alias{.alignment_read}
\title{Read alignments}
\usage{
.alignment_read(flpth, format)
}
\arguments{
\item{flpth}{File path to output file. Character.}

\item{format}{Sequence file format. Character.}
}
\value{
alignment
}
\description{
Return "alignment" from a filepath.
}
\details{
If \code{format} is 'guess', file format will be determined from
file extension.
}
\seealso{
\link{\code{seqinr::read.alignment}}
}
