% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{.sequences_write}
\alias{.sequences_write}
\title{Write sequences}
\usage{
.sequences_write(flpth, x, width)
}
\arguments{
\item{flpth}{File path to output file. Character.}

\item{x}{Sequence object, either "alignment" or "supermatrix".}

\item{width}{Number of sequence columns. Integer.}
}
\value{
Logical
}
\description{
Return "alignment" or "supermatrix"  object to .FASTA file.
}
