% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{drop_tips}
\alias{drop_tips}
\title{Drop tips from a supermatrix}
\usage{
drop_tips(supermatrix, cutoff = 0.5)
}
\arguments{
\item{supermatrix}{supermatrix object}

\item{cutoff}{Proportion of non-gap cutoff}
}
\value{
supermatrix
}
\description{
Return a supermatrix object with tips dropped that have more
non-gaps than cutoff.
}
