% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{format_guess}
\alias{format_guess}
\title{Guess the sequence file format}
\usage{
format_guess(flpth)
}
\arguments{
\item{flpth}{File path to sequence file. Character.}
}
\value{
character
}
\description{
Return sequence file format from a file path based on the
file's extension. If no format found, errors.
}
