% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaius.R
\docType{package}
\name{gaius}
\alias{gaius}
\alias{gaius-package}
\title{gaius: Divide and conquer phylogenetic super -matrices and -trees in R}
\description{
For more information visit the gaius website
(\url{https://antonellilab.github.io/gaius/}).
}
