% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{gapmatrix_get}
\alias{gapmatrix_get}
\alias{pmissing.gapmatrix}
\title{Generate a gaps matrix}
\usage{
gapmatrix_get(alignment)

\method{pmissing}{gapmatrix}(x)
}
\arguments{
\item{alignment}{alignment object}

\item{x}{gapmatrix object}
}
\value{
gapmatrix
}
\description{
Return a matrix of TRUE and FALSE corresponding to '-' in a
alignment.
}
