% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{msar_wrapper}
\alias{msar_wrapper}
\title{Visualise sequences}
\usage{
msar_wrapper(sequences, nms)
}
\arguments{
\item{sequences}{list of biological sequences as separate characters}

\item{nms}{names of sequences in list}
}
\description{
Plot a Shiny multiple alignment object.
}
