% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{nms_from_fasta}
\alias{nms_from_fasta}
\title{Read tip labels from sequence file}
\usage{
nms_from_fasta(flpth)
}
\arguments{
\item{flpth}{File path to .fasta file}
}
\value{
character vector
}
\description{
Return .
}
