% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\alias{default_pset}
\alias{default_pget}
\alias{pset}
\alias{pget}
\title{Set and get parameters}
\usage{
default_pset()

default_pget()

pset(val, parameter = names(default_pget()))

pget(parameter = names(default_pget()))
}
\arguments{
\item{val}{Value to be assigned to parameter}

\item{parameter}{Name of parameter}
}
\description{
Set new and default parameters. Look up currently set
parameters.
}
