% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{partition_file}
\alias{partition_file}
\title{Generate partition file for RAxML}
\usage{
partition_file(nbp, type = "DNA, gene", flpth)
}
\arguments{
\item{nbp}{Number of base pairs per cluster, character vector}

\item{type}{Type of cluster, e.g. DNA, gene, character vector}

\item{flpth}{File path for resulting file}
}
\description{
Write out partition text file for RAxML from supermatrix
information.
}
