% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{phylo}
\alias{phylo}
\title{Generate phylogenies from supermatrices}
\usage{
phylo(supermatrices, method)
}
\arguments{
\item{supermatrices}{supermatrices object}

\item{method}{phylogeny generation method}
}
\value{
filepaths to phylogenies, character vector
}
\description{
Create phylogenies from supermatrices using method specified.
}
