% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sequences_filter}
\alias{sequences_filter}
\title{Filter sequecnes from alignments}
\usage{
sequences_filter(alignment_list, cutoff = 0.9, min_nbps = 200)
}
\value{
alignment_list
}
\description{
Return a filtered alignment. Columns will be dropped
that have fewer proportion of non-gaps then cutoff and genes will be dropped
that are less than min_nbps long.
}
