% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sequences_select}
\alias{sequences_select}
\title{Select sequences from alignments}
\usage{
sequences_select(alignment_list, nms)
}
\arguments{
\item{alignment_list}{alignment list object}

\item{nms}{names of tips to be selected}
}
\value{
alignment_list
}
\description{
Return a alignment of just sequences corresponding to
nms.
}
