% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{sequences_write}
\alias{sequences_write}
\title{Write sequences to file}
\usage{
sequences_write(x, flpth, width = 80)
}
\arguments{
\item{x}{sequences as "supermatrix", "supermatrices", "alignment" or
"alignment_list" objects.}

\item{flpth}{File path to output file or folder. Character.}

\item{width}{Max number of base pairs per line, default 80. Integer.}
}
\value{
Logical
}
\description{
Takes a sequence-based object and writes the sequences to file
in .FASTA format. For objects with multiple sequence groups, "supermatrices"
and "alignment_list", the \code{flpth} should point to a directory not a
file. Files are then created in this directory based on the group name for
each sequence group.
}
