\name{supermatrices}
\alias{supermatrices}
\docType{data}
\title{
Example supermatrices
}
\description{
Example supermatrices object for demonstration purposes.
}
\usage{data("supermatrices")}
\format{
  A large matrix object containing fake DNA sequence.
}
\examples{
data(supermatrices)
print(supermatrices)
}
\keyword{datasets}