\name{supermatrix}
\alias{supermatrix}
\docType{data}
\title{
Example supermatrix
}
\description{
Example supermatrix object for demonstration purposes.
}
\usage{data("supermatrix")}
\format{
  A large matrix object containing fake DNA sequence.
}
\examples{
data(supermatrix)
print(supermatrix)
}
\keyword{datasets}