% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{supermatrix_get}
\alias{supermatrix_get}
\title{Convert list of alignments into a supermatrix}
\usage{
supermatrix_get(alignment_list)
}
\value{
supermatrix
}
\description{
Return a supermatrix object from a list of alignments. The
supermatrix object is smaller.
}
