% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertree.R
\name{supertree}
\alias{supertree}
\title{Create supertree}
\usage{
supertree(specieslevel, backbone)
}
\arguments{
\item{specieslevel}{file path to species-level phylogen(ies/y)}

\item{backbone}{file path to backbone phylogeny}
}
\value{
filepath to phylogeny, character
}
\description{
Generate supertree by combining species-level trees to backbone
tree.
}
