% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz}
\alias{viz}
\alias{viz.alignment}
\alias{viz.supermatrix}
\title{Visualise an alignment or supermatrix}
\usage{
\method{viz}{alignment}(x)

\method{viz}{supermatrix}(x)
}
\arguments{
\item{x}{alignment object}
}
\description{
Plot a Shiny multiple alignment object.
}
