% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{alignment_read}
\alias{alignment_read}
\title{Read in sequence alignment(s)}
\usage{
alignment_read(flpths)
}
\arguments{
\item{flpths}{File path(s) to .fasta file(s)}
}
\value{
alignment list
}
\description{
Return alignment list object from reading in sequences from a
single FASTA file or multiple FASTA files.
}
